+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.21.3
+
+* fixed invalidate for CPManager