+++ CHANGELOG.md
@@ -1,4 +1,2 @@
-# v2.21.3
+# v2.21.4
 
-* fixed invalidate for CPManager
-* exported max_number for CPManager